refactored socials
