+++ src/components/Footer/index.css
@@ -7,14 +7,3 @@
     align-items: center;
 }
 
-.socials {
-    display: flex;
-    padding-left: 5%;
-}
-
-.icon {
-    padding-left: 20px;
-    padding-right: 20px;
-}
-
-
