+++ src/components/Footer/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Socials from './Socials';
+import Socials from '../Socials';
 import './index.css';
 
 const Footer = (props) => {
