+++ src/components/Socials/index.css
@@ -1,0 +1,11 @@
+.socials {
+    display: flex;
+    padding-left: 5%;
+}
+
+.icon {
+    padding-left: 20px;
+    padding-right: 20px;
+}
+
+
